requirements.txt
adal==1.2.2
alabaster==0.7.12
alembic==1.4.2
amqp==2.5.2
analytics-python==1.2.9
ansiwrap==0.8.4
apache-airflow==1.10.10
apispec==1.3.3
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.11.1
asn1crypto==1.3.0
astroid==2.3.3
async-generator==1.10
atlasclient==1.0.0
attrs==19.3.0
aws-sam-translator==1.22.0
aws-xray-sdk==2.5.0
azure-common==1.1.25
azure-cosmos==3.1.2
azure-datalake-store==0.0.48
azure-mgmt-containerinstance==1.5.0
azure-mgmt-resource==9.0.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.6.3.0
black==19.10b0
bleach==3.1.4
blinker==1.4
boto==2.49.0
boto3==1.12.46
botocore==1.14.17
cached-property==1.5.1
cachetools==4.1.0
cassandra-driver==3.20.2
cattrs==0.9.2
celery==4.4.2
certifi==2020.4.5.1
cffi==1.13.2
cfgv==3.1.0
cfn-lint==0.29.6
cgroupspy==0.1.6
chardet==3.0.4
click==6.7
cloudant==0.5.10
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
coverage==5.1
croniter==0.3.31
cryptography==2.9.2
cx-Oracle==7.3.0
Cython==0.29.16
datadog==0.35.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
dnspython==1.16.0
docker==3.7.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
ecdsa==0.15
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
entrypoints==0.3
fastavro==0.23.2
filelock==3.0.12
flake8==3.7.9
flake8-colors==0.1.6
flaky==3.6.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.2
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
flower==0.9.4
freezegun==0.3.15
fsspec==0.7.3
funcsigs==1.0.2
future==0.18.2
future-fstrings==1.2.0
gcsfs==0.6.1
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.2.1
google-cloud-container==0.5.0
google-cloud-core==1.3.0
google-cloud-dlp==0.14.0
google-cloud-language==1.3.0
google-cloud-secret-manager==0.2.0
google-cloud-spanner==1.15.1
google-cloud-speech==1.3.2
google-cloud-storage==1.28.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==2.0.1
google-cloud-videointelligence==1.14.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.14
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
grpcio-gcp==0.2.2
gunicorn==19.10.0
hdfs==2.5.8
hmsclient==0.1.1
httplib2==0.17.3
humanize==0.5.1
hvac==0.10.1
identify==1.4.15
idna==2.9
ijson==2.6.1
imagesize==1.2.0
importlib-metadata==1.6.0
inflection==0.4.0
ipdb==0.13.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
JayDeBeApi==1.1.1
jedi==0.17.0
Jinja2==2.11.2
jira==2.0.0
jmespath==0.9.5
JPype1==0.7.1
json-merge-patch==0.2
jsondiff==1.1.2
jsonpatch==1.25
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==4.6.8
kubernetes==11.0.0
lazy-object-proxy==1.4.3
ldap3==2.7
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
mongomock==3.19.0
more-itertools==8.2.0
moto==1.3.14
msrest==0.6.13
msrestazure==0.6.3
multi-key-dict==2.0.3
mypy==0.720
mypy-extensions==0.4.3
mysqlclient==1.3.14
nbclient==0.2.0
nbconvert==5.6.1
nbformat==5.0.6
nest-asyncio==1.3.2
networkx==2.4
nodeenv==1.3.5
notebook==6.0.3
nteract-scrapbook==0.4.1
ntlm-auth==1.4.0
numpy==1.18.1
oauthlib==3.1.0
oscrypto==1.2.0
packaging==20.3
pandas==1.0.3
pandas-gbq==0.13.1
pandocfilters==1.4.2
papermill==2.1.1
parameterized==0.7.4
paramiko==2.7.1
parso==0.7.0
pathspec==0.8.0
pbr==5.4.5
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
pinotdb==0.1.1
pluggy==0.13.1
pre-commit==2.3.0
presto-python-client==0.7.0
prison==0.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodomex==3.9.7
pydata-google-auth==1.1.0
pydruid==0.5.8
pyflakes==2.1.1
Pygments==2.6.1
PyHive==0.6.2
PyJWT==1.7.1
pykerberos==1.2.1
pymongo==3.10.1
pymssql==2.1.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pysftp==0.2.9
PySmbClient==0.1.5
pyspark==2.4.5
pytest==5.4.1
pytest-cov==2.8.1
pytest-instafail==0.4.1.post0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.2.7
python-jenkins==1.7.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
pywinrm==0.4.1
PyYAML==5.3.1
pyzmq==19.0.0
qds-sdk==1.15.2
qtconsole==4.7.3
QtPy==1.9.0
redis==3.4.1
regex==2020.4.4
requests==2.23.0
requests-futures==0.9.4
requests-kerberos==0.12.0
requests-mock==1.7.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.10.14
rsa==4.0
s3transfer==0.3.3
sasl==0.2.1
Send2Trash==1.5.0
sendgrid==5.6.0
sentinels==1.0.0
sentry-sdk==0.14.3
setproctitle==1.1.10
six==1.14.0
slackclient==1.3.2
snowballstemmer==2.0.0
snowflake-connector-python==2.2.4
snowflake-sqlalchemy==1.2.3
soupsieve==2.0
Sphinx==3.0.2
sphinx-argparse==0.2.5
sphinx-autoapi==1.0.0
sphinx-jinja==1.1.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.16
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.3
sshpubkeys==3.1.0
sshtunnel==0.1.5
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
textwrap3==0.9.2
thrift==0.13.0
thrift-sasl==0.4.2
toml==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
vertica-python==0.10.3
vine==1.3.0
virtualenv==20.0.18
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.3.1
xmltodict==0.12.0
yamllint==1.23.0
zdesk==2.7.1
zipp==3.1.0
zope.deprecation==4.4.0
